% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarymissing.R
\name{summarymissing}
\alias{summarymissing}
\title{Missingness Summary.}
\usage{
summarymissing(data, sum.col = TRUE, missgraph = FALSE)
}
\arguments{
\item{data}{Data.frame input.}

\item{sum.col}{Logical. Returns the names of missing columns. Defaults to TRUE.}

\item{missgraph}{Logical. Plots a graph of the overall missing summary. Defaults to TRUE.}
}
\value{
The missing percentages of the overall data.frame, by rows and by columns.
}
\description{
This function returns a missingness summary of a data.frame.
}
\examples{
summarymissing(testdf, sum.col = TRUE, missgraph = FALSE)
}
\keyword{missing,}
\keyword{summary}
