% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdf}
\alias{testdf}
\title{Test data set for summarymissing.}
\format{A data frame with 4 rows and 3 variables:
\describe{
  \item{Age}{in years}
  \item{Gender}{M / F}
  \item{Happiness}{0 to 10}

}}
\usage{
testdf
}
\description{
A sample data set to showcase the function summary missing
}
\keyword{datasets}
